Group install_ci tests under cli/, misc/, and nvbugs/ subdirs to prepare for systematic coverage
