+++ source/isaaclab/test/install_ci/cli/cli_install_training_conda_env.test.py
@@ -1,0 +1,116 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""End-to-end installation and training workflow tests (conda).
+
+Covers conda-based installation paths:
+  - conda × kitless (core-only, ``-i none``)
+  - conda × newton training (``-i newton,rl[rsl-rl]``)
+
+Tests in this file are intentionally slow and GPU-dependent.  They are
+gated behind pytest markers so they only run in the appropriate CI
+environment:
+
+  ``@pytest.mark.conda`` – routed to the conda-enabled Docker image
+  ``@pytest.mark.gpu``   – requires a GPU
+  ``@pytest.mark.slow``  – skipped in fast/smoke runs
+"""
+
+from __future__ import annotations
+
+import shutil
+
+import pytest
+from utils import Conda_Mixin
+
+# ---------------------------------------------------------------------------
+# Shared training helper
+# ---------------------------------------------------------------------------
+
+_TRAIN_CMD = [
+    "train",
+    "--rl_library",
+    "rsl_rl",
+    "--task",
+    "Isaac-Cartpole-Direct-v0",
+    "--num_envs",
+    "16",
+    "presets=newton_mjwarp",
+    "--max_iterations",
+    "5",
+    "--headless",
+]
+
+
+def _assert_training_passed(result) -> None:
+    output = result.stdout + (result.stderr or "")
+    assert result.returncode == 0, f"Training failed (rc={result.returncode}):\n{output}"
+    assert "Traceback (most recent call last):" not in output, f"Training produced a traceback:\n{output}"
+    assert "Training time:" in output, f"Training did not report completion:\n{output}"
+
+
+# ---------------------------------------------------------------------------
+# conda-based tests
+# ---------------------------------------------------------------------------
+
+
+class TestCondaWorkflow(Conda_Mixin):
+    """Installation and training smoke tests using conda environments."""
+
+    @classmethod
+    def setup_class(cls):
+        if not shutil.which("conda"):
+            pytest.skip("conda is not available")
+
+    @pytest.mark.cli
+    @pytest.mark.conda
+    @pytest.mark.slow
+    @pytest.mark.gpu
+    @pytest.mark.timeout(1200)
+    def test_conda_none_installs_core_submodules(self, isaaclab_root):
+        """conda + ``./isaaclab.sh -i none`` installs all core submodules without extras."""
+        try:
+            self.create_conda_env(isaaclab_root)
+            result = self.run_in_conda_env(
+                [str(self.cli_script), "-i", "none"],
+                cwd=isaaclab_root,
+                timeout=900,
+            )
+            assert result.returncode == 0, f"conda isaaclab -i none failed:\n{result.stdout}\n{result.stderr}"
+            for pkg in ("isaaclab", "isaaclab_assets", "isaaclab_tasks", "isaaclab_physx"):
+                r = self.run_in_conda_env(
+                    [str(self.python), "-c", f"import {pkg}; print({pkg!r}, 'ok')"],
+                    cwd=isaaclab_root,
+                    timeout=60,
+                )
+                assert r.returncode == 0, f"{pkg} not importable after conda -i none:\n{r.stdout}\n{r.stderr}"
+        finally:
+            self.destroy_conda_env()
+
+    @pytest.mark.cli
+    @pytest.mark.conda
+    @pytest.mark.slow
+    @pytest.mark.gpu
+    @pytest.mark.timeout(1800)
+    def test_conda_newton_rsl_rl_trains_cartpole(self, isaaclab_root):
+        """conda + ``./isaaclab.sh -i newton,rl[rsl-rl]`` + training completes successfully."""
+        try:
+            self.create_conda_env(isaaclab_root)
+            result = self.run_in_conda_env(
+                [str(self.cli_script), "-i", "newton,rl[rsl-rl]"],
+                cwd=isaaclab_root,
+                timeout=1200,
+            )
+            assert result.returncode == 0, (
+                f"conda isaaclab -i newton,rl[rsl-rl] failed:\n{result.stdout}\n{result.stderr}"
+            )
+            result = self.run_in_conda_env(
+                [str(self.cli_script)] + _TRAIN_CMD,
+                cwd=isaaclab_root,
+                timeout=600,
+            )
+            _assert_training_passed(result)
+        finally:
+            self.destroy_conda_env()
