+++ source/isaaclab/test/install_ci/cli/cli_install_visualizer_uv_env.test.py
@@ -41,7 +41,6 @@ class Test_Install_Visualizer(UV_Mixin):
 
     @pytest.mark.uv
     @pytest.mark.slow
-    @pytest.mark.native
     @pytest.mark.timeout(1800)
     def test_visualizer_rerun_backend_importable(self, isaaclab_root):
         """rerun-sdk is importable after ./isaaclab.sh -i 'visualizer[rerun]'."""
@@ -60,7 +59,6 @@ class Test_Install_Visualizer(UV_Mixin):
 
     @pytest.mark.uv
     @pytest.mark.slow
-    @pytest.mark.native
     @pytest.mark.timeout(1800)
     def test_visualizer_viser_backend_importable(self, isaaclab_root):
         """viser is importable after ./isaaclab.sh -i 'visualizer[viser]'."""
@@ -79,7 +77,6 @@ class Test_Install_Visualizer(UV_Mixin):
 
     @pytest.mark.uv
     @pytest.mark.slow
-    @pytest.mark.native
     @pytest.mark.timeout(1800)
     def test_visualizer_default_installs_all_backends(self, isaaclab_root):
         """./isaaclab.sh -i visualizer (no selector) installs all visualizer backends."""
@@ -101,7 +98,6 @@ class Test_Install_Visualizer(UV_Mixin):
 
     @pytest.mark.uv
     @pytest.mark.slow
-    @pytest.mark.native
     @pytest.mark.timeout(1800)
     def test_visualizer_all_backends_pull_newton_sim(self, isaaclab_root):
         """Every visualizer backend install also provides the newton package."""
@@ -123,7 +119,6 @@ class Test_Install_Visualizer(UV_Mixin):
     @pytest.mark.uv
     @pytest.mark.gpu
     @pytest.mark.slow
-    @pytest.mark.native
     @pytest.mark.timeout(3600)
     def test_train_with_rerun_visualizer(self, isaaclab_root):
         """Training with --visualizer rerun works after ./isaaclab.sh -i 'newton,rl[rsl-rl],visualizer[rerun]'."""
