+++ source/isaaclab/test/install_ci/pytest.ini
@@ -6,10 +6,13 @@ python_files =
     *_test.py
     *_tests.py
 markers =
-    bug: bug-regression tests (use bug id as argument)
+    bug: regression tests (bug id as argument)
     gpu: tests that require a GPU
-    docker: tests that only run inside Docker
-    native: tests that only run natively (not in Docker)
+    docker: tests that ONLY run inside Docker
+    native: tests that ONLY run natively (not in Docker)
     slow: tests that take a long time
-    uv: tests that require the uv package manager
+    uv: tests that require the uv installed
+    conda: tests that require conda installed
+    cli: tests that run "./isaaclab.sh -i"
 timeout = 1200
+addopts = --import-mode=importlib
